♻️  mission_ nodes
